Crate evmc_vm

Rust bindings for EVMC (Ethereum Client-VM Connector API).

Have a look at evmc-declare to declare an EVMC compatible VM. This crate documents how to use certain data types.

Re-exports

pub use evmc_sys as ffi;

Structs

EvmcContainer
Container struct for EVMC instances and user-defined data.

ExecutionContext
EVMC context structure. Exposes the EVMC host functions, message data, and transaction context to the executing VM.

ExecutionMessage
EVMC execution message structure.

ExecutionResult
EVMC result structure.

Traits

EvmcVm
Trait EVMC VMs have to implement.

Type Definitions

Address
EVMC address

Bytes32
EVMC 32 bytes value (used for hashes)

Capabilities
EVMC VM capabilities.

ExecutionTxContext
EVMC transaction context structure.

MessageFlags
EVMC message (call) flags.

MessageKind
EVMC call kind.

Revision
EVMC VM revision.

StatusCode
EVMC status code.

StorageStatus
EVMC storage status.

Uint256
EVMC big-endian 256-bit integer